requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2022.5.18
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
debugpy==1.6.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jupyter-client==7.3.1
jupyter-core==4.10.0
kombu==5.2.4
Markdown==3.3.7
matplotlib-inline==0.1.3
Naked==0.1.31
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.14.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
redis==4.3.1
requests==2.27.1
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tornado==6.1
traitlets==5.2.1.post0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
